Move Prometheus proxy to own package and properly handle not found tenants
